init: upgraded next-mdx-remote. blog, alternative upgraded.
