+++ apps/www/next.config.mjs
@@ -1,0 +1,68 @@
+import nextMdx from '@next/mdx'
+
+import remarkGfm from 'remark-gfm'
+import rehypeSlug from 'rehype-slug'
+
+import rewrites from './lib/rewrites.js'
+import redirects from './lib/redirects.js'
+
+const withMDX = nextMdx({
+  extension: /\.mdx?$/,
+  options: {
+    remarkPlugins: [remarkGfm],
+    rehypePlugins: [rehypeSlug],
+  },
+})
+
+const nextConfig = {
+  basePath: '',
+  pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'md', 'mdx'],
+  trailingSlash: false,
+  images: {
+    dangerouslyAllowSVG: true,
+    domains: [
+      'avatars.githubusercontent.com',
+      'github.com',
+      'ca.slack-edge.com',
+      'res.cloudinary.com',
+      'images.unsplash.com',
+      'supabase.com',
+      'obuldanrptloktxcffvn.supabase.co',
+      'avatars.githubusercontent.com',
+      'colab.research.google.com',
+      'api.producthunt.com',
+      'https://s3-us-west-2.amazonaws.com',
+      's3-us-west-2.amazonaws.com',
+      'user-images.githubusercontent.com',
+    ],
+  },
+  async headers() {
+    return [
+      {
+        source: '/(.*)',
+        headers: [
+          {
+            key: 'Strict-Transport-Security',
+            value: '',
+          },
+          {
+            key: 'X-Robots-Tag',
+            value: 'all',
+          },
+          {
+            key: 'X-Frame-Options',
+            value: 'DENY',
+          },
+        ],
+      },
+    ]
+  },
+  async rewrites() {
+    return rewrites
+  },
+  async redirects() {
+    return redirects
+  },
+}
+
+export default withMDX(nextConfig)
