+++ apps/www/package.json
@@ -28,7 +28,7 @@
     "markdown-toc": "^1.2.0",
     "marked": "^4.0.12",
     "next": "latest",
-    "next-mdx-remote": "^2.1.3",
+    "next-mdx-remote": "^4.1.0",
     "next-seo": "^4.20.0",
     "nodemailer": "^6.7.3",
     "parse-numeric-range": "^1.3.0",
@@ -40,10 +40,10 @@
     "react-syntax-highlighter": "^15.5.0",
     "react-tooltip": "^4.2.17",
     "react-transition-group": "^4.4.1",
-    "rehype-slug": "^4.0.1",
     "remark": "^13.0.0",
-    "remark-gfm": "^1.0.0",
-    "remark-html": "10.0.0",
+    "rehype-slug": "^5.0.1",
+    "remark-gfm": "^3.0.1",
+    "remark-html": "^15.0.1",
     "swiper": "^6.5.0",
     "swr": "^0.5.4",
     "typescript": "^4.0.3",
